refactor: organize DDD layers
