+++ src/application/README.md
@@ -1,0 +1,3 @@
+# Application Layer
+
+Place use-case services coordinating domain models and infrastructure here.
