+++ src/domain/pokemon.ts
